update visibility value to floats

+++ datamodel/datamodel.go
@@ -65,9 +65,9 @@ type Response struct {
 			QualityControl string      `json:"qualityControl"`
 		} `json:"seaLevelPressure"`
 		Visibility struct {
-			Value          int    `json:"value"`
-			UnitCode       string `json:"unitCode"`
-			QualityControl string `json:"qualityControl"`
+			Value          float64 `json:"value"`
+			UnitCode       string  `json:"unitCode"`
+			QualityControl string  `json:"qualityControl"`
 		} `json:"visibility"`
 		MaxTemperatureLast24Hours struct {
 			Value          interface{} `json:"value"`
